+++ adapter/moloco/src/main/java/org/bidon/moloco/ext/Ext.kt
@@ -1,8 +1,10 @@
 package org.bidon.moloco.ext
 
 import com.moloco.sdk.publisher.Banner
+import com.moloco.sdk.publisher.MediationInfo
 import com.moloco.sdk.publisher.Moloco
 import com.moloco.sdk.publisher.MolocoAdError
+import org.bidon.moloco.EMPTY_MEDIATOR
 import org.bidon.moloco.EMPTY_WATERMARK
 import org.bidon.moloco.MolocoDemandId
 import org.bidon.sdk.ads.banner.BannerFormat
@@ -42,20 +44,35 @@ internal fun Moloco.createBannerAd(
 
     try {
         when (bannerSize) {
-            BannerFormat.Banner ->
-                Moloco.createBanner(adUnitId, EMPTY_WATERMARK, sdkCallback)
+            BannerFormat.Banner -> {
+                Moloco.createBanner(
+                    mediationInfo = MediationInfo(EMPTY_MEDIATOR),
+                    adUnitId = adUnitId,
+                    watermarkString = EMPTY_WATERMARK,
+                    callback = sdkCallback
+                )
+            }
 
-            BannerFormat.LeaderBoard ->
-                Moloco.createBannerTablet(adUnitId, EMPTY_WATERMARK, sdkCallback)
+            BannerFormat.LeaderBoard -> {
+                Moloco.createBannerTablet(
+                    mediationInfo = MediationInfo(EMPTY_MEDIATOR),
+                    adUnitId = adUnitId,
+                    watermarkString = EMPTY_WATERMARK,
+                    callback = sdkCallback
+                )
+            }
 
-            BannerFormat.MRec ->
-                Moloco.createMREC(adUnitId, EMPTY_WATERMARK, sdkCallback)
+            BannerFormat.MRec -> {
+                Moloco.createMREC(
+                    mediationInfo = MediationInfo(EMPTY_MEDIATOR),
+                    adUnitId = adUnitId,
+                    watermarkString = EMPTY_WATERMARK,
+                    callback = sdkCallback
+                )
+            }
 
             BannerFormat.Adaptive -> {
-                callback(
-                    null,
-                    IllegalStateException("Adaptive format should have been resolved")
-                )
+                callback(null, IllegalStateException("Adaptive format should have been resolved"))
             }
         }
     } catch (t: Throwable) {