+++ adapter/yandex/src/main/java/org/bidon/yandex/impl/YandexRewardedImpl.kt
@@ -19,6 +19,7 @@ import org.bidon.sdk.config.BidonError
 import org.bidon.sdk.logs.logging.impl.logInfo
 import org.bidon.sdk.stats.StatisticsCollector
 import org.bidon.sdk.stats.impl.StatisticsCollectorImpl
+import org.bidon.sdk.stats.models.BidType
 import org.bidon.yandex.ext.asBidonAdValue
 import org.bidon.yandex.ext.asBidonError
 
@@ -46,7 +47,20 @@ internal class YandexRewardedImpl :
         val adUnitId = adParams.adUnitId
             ?: return emitEvent(AdEvent.LoadFailed(BidonError.IncorrectAdUnit(demandId = demandId, message = "adUnitId")))
 
-        val adRequestConfiguration = AdRequestConfiguration.Builder(adUnitId).build()
+        val adRequestConfiguration = AdRequestConfiguration.Builder(adUnitId)
+            .apply {
+                when (adParams.adUnit.bidType) {
+                    BidType.RTB -> {
+                        val signalData = adParams.signalData
+                            ?: return emitEvent(AdEvent.LoadFailed(BidonError.IncorrectAdUnit(demandId = demandId, message = "signalData")))
+                        setBiddingData(signalData)
+                    }
+
+                    BidType.CPM -> Unit
+                }
+            }
+            .build()
+
         val adLoadListener = object : RewardedAdLoadListener {
             override fun onAdLoaded(rewarded: RewardedAd) {
                 this@YandexRewardedImpl.rewardedAd = rewarded