+++ bidon/src/main/java/org/bidon/sdk/ads/interstitial/InterstitialImpl.kt
@@ -152,7 +152,10 @@ internal class InterstitialImpl(
             logInfo(TAG, "Sdk is not initialized")
             return
         }
+        logInfo(TAG, "Destroy ad")
         scope.launch(Dispatchers.Main.immediate) {
+            winner?.destroy()
+            winner = null
             adCache.clear()
             observeCallbacksJob?.cancel()
             observeCallbacksJob = null