+++ bidon/src/main/java/org/bidon/sdk/ads/rewarded/RewardedImpl.kt
@@ -150,7 +150,10 @@ internal class RewardedImpl(
             logInfo(TAG, "Sdk is not initialized")
             return
         }
+        logInfo(TAG, "Destroy ad")
         scope.launch(Dispatchers.Main.immediate) {
+            winner?.destroy()
+            winner = null
             adCache.clear()
             observeCallbacksJob?.cancel()
             observeCallbacksJob = null