+++ thirdPartyMediationAdapters/applovin_max/src/main/java/com/applovin/mediation/adapters/ext/BundleExt.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 
 internal fun Bundle.getAsDouble(key: String, defaultValue: Double = 0.0): Double {
     return try {
+        @Suppress("DEPRECATION")
         val value = this.get(key)
         if (value is Number) value.toDouble() else defaultValue
     } catch (_: Exception) {