+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.13.0 (2025.12.04)
+## Features:
+- [BDN-1081](https://appodeal.atlassian.net/browse/BDN-1081) Fixed incorrect bid status marked as SUCCESSFUL when bid.price < pricefloor
+- [BDN-1095](https://appodeal.atlassian.net/browse/BDN-1095) Added proper destroy for ad instances in cache and auction losers to prevent memory leaks
+
 # 0.12.0 (2025.11.07)
 ## Features:
 - [BDN-1011](https://appodeal.atlassian.net/browse/BDN-1011) Decoupled adapter version from sdk core version